Implement automatic provisioning profile generation (macOS hosts only)
If no provisioning profile is provided and the target platform requires
provisioning profiles, Swift Bundler searches the standard Xcode
location for any existing provisioning profiles that match the bundle
identifier, codesigning identity and target device of the current build.
If no matches are found, Swift Bundler generates a dummy Xcode project
and runs a build with `-allowProvisioningUpdates` to trick xcodebuild
into generating the provisioning profile that we require.

This commit also moves back over to upstream XcodeProj and XcodeGen.