Run apps with devicectl when available (Xcode 15 broke ios-deploy)
devicectl is Apple's first-party version of ios-deploy and it seems to
work much better with newer iOS devices. It was introduced in Xcode 15
so we still fallback on ios-deploy when devicectl can't be found.

The thing that broke with ios-deploy was launching apps (so you could
install but couldn't launch automatically with an attached console).